feat: change structure as is needed by Fluentd
