+++ .rubocop.yml
@@ -23,9 +23,9 @@
 Metrics/LineLength:
   Max: 120
 Metrics/MethodLength:
-  Max: 30
+  Max: 40
 Metrics/ClassLength:
-  Max: 300
+  Max: 350
 Metrics/AbcSize:
   Max: 30
 Metrics/CyclomaticComplexity:
